refactor: move over process files
